+++ tests/test_websocket_writer.py
@@ -169,7 +169,8 @@ async def test_concurrent_messages(
         result, _ = reader.feed_data(call_bytes)
         assert result is False
         msg = await queue.read()
-        bytes_data: bytes = msg.data
+        assert msg.type is WSMsgType.BINARY
+        bytes_data = msg.data
         first_char = bytes_data[0:1]
         char_val = ord(first_char)
         assert len(bytes_data) == char_val